Move slot table types into pointer_flow
